# docs/api/openapi/auth-v1.yaml
...
openapi: 3.1.0
info:
  title: Pet Care Auth API v1
  version: 1.0.0
  description: >-
    Strictly derived from docs/ (01 operations, 02 RULE-01/02, 03 FSM-1 Account,
    04 glossary+D-04, 05 domain 4.1, 06 ERD accounts/otps/users).
    Base URL/versioning (Q1) is TBD: no `servers` is declared
    (defaults to "/") and paths below are PROPOSED. Token TTLs (Q4), password
    policy (Q5) intentionally absent. Forgot/reset password added 2026-09-16
    (Q12 DECIDED); /auth/refresh added 2026-09-14 (Q3 DECIDED). Error envelope follows
    docs/convention/backend/04-exception-handling
    ({success, errorCode, message, statusCode, timestamp, traceId}); 410/423/429
    are PROPOSED extensions not yet in that convention. LOCKED has no
    auto-unlock: LOCKED->ACTIVE only via manual UnlockAccount (FSM-1
    invariant #5).
    **Sửa 2026-09-13 (RULE-01-10):** danh tính chính đổi từ `phone` sang
    `email` cho Register/VerifyOtp/ResendOtp — xem Decision Log
    docs/02-business-rules.md mục 01 (cuối). `phone` chuyển thành optional.
...
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh (làm mới token pair)
      description: >-
        IMPLEMENTED 2026-09-14. Rotate refresh token (1-lần-dùng, ADR-0001);
        không có RULE-ID/FSM riêng — session/token layer infra giống Logout.
        Re-check AccountStatus tại thời điểm gọi (RULE-01-01/07).
      operationId: refresh
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token pair mới; refresh token cũ đã bị revoke.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Refresh token không hợp lệ — malformed/hết hạn/sai type/không tìm thấy hoặc đã bị revoke (reuse).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Account không còn ACTIVE (PENDING_VERIFICATION/DEACTIVATED) kể từ lúc refresh token được cấp.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '423':
          description: Account đã bị LOCKED kể từ lúc refresh token được cấp.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: ASSUMPTION A2 — docs only confirm JWT/Session Token (RULE-01-01).
  schemas:
...
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    RefreshTokenResponse:
      type: object
      required: [accessToken, refreshToken, tokenType, expiresIn]
      properties:
        accessToken:
          type: string
          description: Access token mới.
        refreshToken:
          type: string
          description: Refresh token mới (rotate — token cũ dùng trong request đã bị revoke, ADR-0001).
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          format: int64
          description: TTL access token tính bằng giây (ADR-0002, mặc định 900 = 15 phút).
...
    Error:
      type: object
      required: [success, errorCode, message, statusCode, timestamp]
      properties:
        success:
          type: boolean
          example: false
        errorCode:
          type: string
          description: Only ACCESS_DENIED_SCOPE_MISMATCH is CONFIRMED (RULE-02-01, 403); VALIDATION_FAILED/BUSINESS_RULE_VIOLATION/INVALID_STATE_TRANSITION/RESOURCE_NOT_FOUND/CONCURRENCY_CONFLICT follow docs/convention/backend/04-exception-handling; all other codes in ../auth-v1.md section C are PROPOSED (410/423/429 have no convention mapping yet).
        message:
          type: string
        statusCode:
          type: integer
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
        details:
          type: object
          additionalProperties: true
  responses:
    BadRequest:
      description: Validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...